feat: implement pagination, add activity types and stats views; update category management and localization
